pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-ai"
version = "1.0.0"
description = "Multi-Agent Counter-UAS Swarm Tracking Platform"
requires-python = ">=3.11"

dependencies = [
    # Core algorithms
    "numpy>=1.26",
    "scipy>=1.12",
    "filterpy>=1.4",
    "numba>=0.59",

    # Machine learning
    "torch>=2.2",
    "torch-geometric>=2.5",
    "stable-baselines3>=2.3",

    # Data / messaging
    "pydantic>=2.6",
    "redis>=5.0",
    "pyzmq>=26.0",
    "msgpack>=1.0",
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "celery>=5.3",

    # API / web
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "httpx>=0.27",

    # UI
    "streamlit>=1.35",
    "plotly>=5.20",
    "pyvis>=0.3",
    "networkx>=3.2",
    "pandas>=2.2",

    # Sensor (optional hardware)
    "pyrtlsdr>=0.3; extra == 'hardware'",

    # Vision (optional)
    "ultralytics>=8.1; extra == 'vision'",

    # Utils
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-benchmark>=4.0",
    "hypothesis>=6.100",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.9",
    "pre-commit>=3.7",
    "ipython>=8.0",
]
hardware = ["pyrtlsdr>=0.3"]
vision   = ["ultralytics>=8.1"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "services*", "ui*"]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "C4", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.per-file-ignores]
"tests/*"          = ["ANN", "S101"]
"ui/*"             = ["ANN", "E402"]
"core/tracking/*"  = ["N803", "N806"]
"core/swarm/*"     = ["N806"]
"core/agents/*"    = ["N806"]

# ── MyPy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["tests/", "ui/", "scripts/"]

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with -m 'not slow')",
    "integration: requires running Docker services",
    "scenario: full end-to-end scenario test",
    "benchmark: performance regression test",
]
filterwarnings = [
    "error::DeprecationWarning",
    "ignore::UserWarning:torch",
]

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source   = ["core", "services"]
omit     = ["*/tests/*", "*/migrations/*", "*/scripts/*"]
branch   = true

[tool.coverage.report]
fail_under   = 80
show_missing = true
skip_covered = false
